requirements.txt
fastapi[standard]
SQLAlchemy>=2.0
google-adk
uvicorn[standard]